requirements.txt
altgraph==0.17.4
black==24.10.0
cffi==1.17.1
click==8.1.8
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.1.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyYAML==6.0.2
setuptools==74.1.2
tcod==16.2.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
